Development: Code Restructure
Description:
1- Moved unused code to archive folder.. To be deleted in future

Info:
1- Using VSCode for development on macOS 13.5
