+++ archive/bad_pixel_interpolator.cpp
